+++ src/executor/builtin_cmd.c
@@ -1,0 +1,8 @@
+#include "minishell.h"
+
+void	builtin_command(t_minivault	*minivault, t_command *command)
+{
+	// run builtin commands here
+	(void) command;
+	(void) minivault;
+}
