+++ src/executor/executor.c
@@ -1,4 +1,3 @@
-
 #include "minishell.h"
 
 bool	is_cmd_builtin(char **builtin_list, char *cmd)
@@ -11,107 +10,6 @@ bool	is_cmd_builtin(char **builtin_list, char *cmd)
 			return (true);
 	}
 	return (false);
-}
-
-char	*get_command_path(char *path_list, char *command)
-{
-	char *temp;
-	int i = 0;
-	if (!path_list)
-	{
-		//path_list is empty deal with it
-		return ;
-	}
-	while (path_list[i])
-	{
-		temp = ft_strjoin(path_list[i], "/");
-		temp = ft_strjoin(temp, command);
-		if (access(temp, F_OK) == 0) // returns if the file exists in the path list
-			return (temp);
-		free(temp);
-		i++;
-	}
-	// command doesnt exit
-}
-
-void	builtin_command(t_minivault	*minivault, t_command *command)
-{
-	// run builtin commands here
-	(void) command;
-	(void) minivault;
-}
-
-size_t	get_list_size(t_word *head)
-{
-	t_word *current;
-	size_t size = 0;
-
-	if (!head)
-		return (size);
-	current = head;
-	while (current)
-	{
-		size++;
-		current = current->next;
-	}
-	return (size);
-}
-
-char	**get_arguments(t_word *words)
-{
-	t_word	*current;
-	int	i = 0;
-	char	**arguments = NULL;
-
-	if (!words)
-		return (NULL); // list is empty something is fucked
-	if (words->next == NULL)
-		return (NULL); // single command with no arguments
-	arguments = (char**)malloc(get_list_size(words) * sizeof(char *));
-	if (!arguments)
-		return (NULL);
-	current = words->next; // to skip the command name wich is the first node
-	while (current)
-	{
-        arguments[i++] = strdup(current->word);
-        current = current->next;
-    }
-	arguments[i] = NULL;
-	return (arguments);
-}
-void	system_command(t_minivault	*minivault, t_command *command)
-{
-	pid_t	child;
-	char	*cmd_path;
-
-	cmd_path = get_command_path(minivault->path, command->words);
-	if (!cmd_path)
-	{
-		// path list is fucked or the cmd doesnt exist
-		return ;
-	}
-	child = fork();
-	if (child == -1)
-	{
-		// fork failed
-		// deal with it
-		return ;
-	}
-	if (child == 0)
-	{
-			close(minivault->pipe_fd[0]); // close the read end of the pipe
-			dup2(minivault->pipe_fd[1], STDOUT_FILENO); //redirect stdout to the write end of the pipe
-			// dup2(infiles, STDIN_FILENO);
-			// TODO: create 2d array with the command and the options
-			if (execve(cmd_path, get_arguments(command->words), minivault->env_list) == -1)
-			{
-				// execve failed 
-				// deal with the error
-				return ;
-			}
-	}
-
-
 }
 
 // TODO: deal with redirections
@@ -139,74 +37,13 @@ void	execute_command(t_minivault *minivault, t_command *command)
 
 void	executor(t_minivault *minivault, t_command **pipeline)
 {
-	int i = -1;
+	int p_count = -1;
+	int i = 0;
 
-	while (pipeline[++i])
-		execute_command(minivault, pipeline[i]);
+	while (pipeline[++p_count])
+		execute_command(minivault, pipeline[p_count]);
+	close(minivault->pipe_fd[0]);
+	close(minivault->pipe_fd[1]);
+
 }
 
-/*
-void	child1(void)
-{
-	pid_t	child_1;
-
-	child_1 = fork();
-	if (child_1 == -1)
-	{
-		perror("In child 1 fork error");
-		exit(EXIT_FAILURE);
-	}
-	else if (child_1 == 0)
-	{
-		open_infile();
-    	close(data()->pipe_fd[0]);
-    	dup2(data()->pipe_fd[1], STDOUT_FILENO);  // Redirect stdout to the write end of the pipe
-		dup2(data()->infile_fd, STDIN_FILENO);
-    	if (execve(data()->cmd1_path, ft_split(data()->cmd1, ' '), args()->env) == -1) 
-		{
-    	    perror("Error executing command1");
-    	    exit(EXIT_FAILURE); // Explicitly exit the child process on error
-    	}
-	}
-}
-
-void	child2(void) 
-{
-	pid_t	child_2;
-
-	child_2 = fork();
-
-	if (child_2 == -1)
-	{
-		perror("In child 2 fork error");
-		exit(EXIT_FAILURE);
-	}
-	else if (child_2 == 0)
-	{
-		open_outfile();
-		close(data()->pipe_fd[1]);
-    	dup2(data()->pipe_fd[0], STDIN_FILENO);  // Redirect stdin to the read end of the pipe
-		dup2(data()->outfile_fd, STDOUT_FILENO);
-    	if (execve(data()->cmd2_path, ft_split(data()->cmd2, ' '), args()->env) == -1)
-		{
-    		perror("Error executing command2");
-  	    	exit(EXIT_FAILURE); // Explicitly exit the child process on error
-    	}
-	}
-}
-
-void	pipex(void)
-{
-	if (pipe(data()->pipe_fd) < 0)
-	{
-		printf("Error: %s\n", strerror(errno));
-		exit(1);
-	}
-	child1();
-	child2();
-	close(data()->pipe_fd[0]);
-	close(data()->pipe_fd[1]);
-	wait(NULL);
-	wait(NULL);
-}
-*/
